extract server.go into package server
