+++ server/server.go
@@ -1,0 +1,6 @@
+package server
+
+
+type Server struct{
+
+}
